+++ nautilus_trader/persistence/catalog/base.py
@@ -21,13 +21,13 @@ from typing import Optional
 from nautilus_trader.model.data.bar import Bar
 from nautilus_trader.model.data.base import DataType
 from nautilus_trader.model.data.base import GenericData
+from nautilus_trader.model.data.book import OrderBookData
 from nautilus_trader.model.data.tick import QuoteTick
 from nautilus_trader.model.data.tick import TradeTick
 from nautilus_trader.model.data.ticker import Ticker
 from nautilus_trader.model.data.venue import InstrumentClose
 from nautilus_trader.model.data.venue import InstrumentStatusUpdate
 from nautilus_trader.model.instruments import Instrument
-from nautilus_trader.model.orderbook.data import OrderBookData
 from nautilus_trader.persistence.external.util import Singleton
 from nautilus_trader.serialization.arrow.util import GENERIC_DATA_PREFIX