+++ tests/integration_tests/adapters/betfair/test_kit.py
@@ -51,9 +51,9 @@ from nautilus_trader.config import ImportableStrategyConfig
 from nautilus_trader.config import LoggingConfig
 from nautilus_trader.config import RiskEngineConfig
 from nautilus_trader.config import StreamingConfig
+from nautilus_trader.model.data.book import OrderBookData
 from nautilus_trader.model.data.tick import TradeTick
 from nautilus_trader.model.instruments.betting import BettingInstrument
-from nautilus_trader.model.orderbook.data import OrderBookData
 from nautilus_trader.persistence.external.readers import LinePreprocessor
 from nautilus_trader.test_kit.stubs.component import TestComponentStubs
 from tests import TESTS_PACKAGE_ROOT