+++ tests/unit_tests/serialization/test_serialization_util.py
@@ -16,8 +16,8 @@
 import pandas as pd
 import pytest
 
+from nautilus_trader.model.data.book import OrderBookData
 from nautilus_trader.model.data.tick import TradeTick
-from nautilus_trader.model.orderbook.data import OrderBookData
 from nautilus_trader.serialization.arrow.util import camel_to_snake_case
 from nautilus_trader.serialization.arrow.util import class_to_filename
 from nautilus_trader.serialization.arrow.util import clean_key